+++ src/Set.php
@@ -76,7 +76,6 @@ final class Set implements \Countable
      *
      * @template V
      * @psalm-pure
-     * @psalm-type RegisterCleanup = callable(callable(): void): void
      *
      * @param callable(RegisterCleanup): \Generator<V> $generator
      *